Add a reminder to be sent 2 days after ad creation
This adds a new reminder e-mail that reminds people
to edit or delete an ad to keep ads relevant.

The rake task is intended to be run by heroku scheduler
and it reminds only 300 oldest emails. That because
the e-mail plan we have supports 400 emails a day
on average and we don't want to eat the quota for other
emails. We have a backlog of unreminded e-mails and
the nighly job will just work through them over several
days.